api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
kafka-python==2.0.2
sqlalchemy==2.0.23